requirements.txt
dash~=1.16.1
plotly~=4.9.0
pandas~=1.1.0
numpy~=1.19.1
matplotlib~=3.3.0
Flask~=1.1.2
Werkzeug~=1.0.1
Jinja2~=3.0.0a1
opencv-python~=4.3.0.36
Pillow~=7.2.0
scipy~=1.4.1
statsmodels~=0.11.1
pytesseract~=0.3.4
future~=0.18.2
torch~=1.5.0
scikit-image~=0.17.2
argparse~=1.4.0
torchvision~=0.6.0
Keras~=2.6.0
six~=1.15.0
lmdb~=0.98
natsort~=7.0.1
seaborn~=0.10.1
sklearn~=0.0
scikit-learn~=0.23.1
deskew~=0.10.3
imutils~=0.5.3
fitz~=0.0.1.dev2
colormap~=1.0.3
pysbd~=0.3.4
gingerit~=0.8.2
